GEO-93: GeoTools/GeoAPI ISO 19115 Metadata synchronization
git-svn-id: https://svn.osgeo.org/geotools/geotools/trunk@24973 e5c1c795-43da-0310-a71f-fac65c449510

+++ gt/modules/extension/openoffice/src/main/java/org/geotools/openoffice/Referencing.java
@@ -666,10 +666,10 @@ public final class Referencing extends Formulas implements XReferencing {
             return getFailure(4,4);
         }
         return new double[][] {
-            new double[] {box.getNorthBoundLatitude(),
-                          box.getWestBoundLongitude()},
-            new double[] {box.getSouthBoundLatitude(),
-                          box.getEastBoundLongitude()}};
+            new double[] {box.getNorthBoundLatitude().doubleValue(),
+                          box.getWestBoundLongitude().doubleValue()},
+            new double[] {box.getSouthBoundLatitude().doubleValue(),
+                          box.getEastBoundLongitude().doubleValue()}};
     }
 
     /**
